Split slack_exporter integration tests out and test main() arg handling
Finding 4: move the full-pipeline test_run_export_* suite into a new
integration-test file test_exporter.py, leaving the genuinely unit-level
_datetime_to_slack_timestamp / _fetch_all_messages tests in exporter_test.py.

Finding 7: extract build_settings_from_args / run_export_or_exit / _build_arg_parser
from main() and unit-test the previously untested CLI boundaries (space-split
channels, --refresh-window-days 0 -> disabled, env TTL parse, error->exit mapping).
